interpreter/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gedl-interpreter"
dynamic = ["version"]
authors = [ 
    {name = "Manuel G. Garcia Alvarez", email = "m.g.garciaalvarez@tudelft.nl"}]

description = "An interpreter for the GEDL language"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
keywords = ["gedl", "interpreter", "language", "GEDSys"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12"
 ]
 dependencies = [
    "python-dotenv >= 1.0.1",
    "shapely >= 2.0.4",
    "requests >=2.31.0",
 ]

[project.optional-dependencies]
dev = [
    "pytest",
    "twine",
    "wheel",
    "hatchling",
    "sphinx",
]

[tool.hatch.version]
path = "src/gedl_interpreter/version.py"
[tool.hatch.build.targets.sdist]
include = [
]
exclude = [
  "docs/",
  "tests/",
  "*.env",
]

[project.urls]
Homepage = "https://github.com/pypa/sampleproject"
Issues = "https://github.com/pypa/sampleproject/issues"